pc/requirements.txt
python 2.7 or higher
numpy
matplotlib
serialpy
argparse
os
time
sys

